feat(scroll padding): implemented brilliant idea of brogrammer nico
we are now using a small 1px rectangle in the bottom right corner
which will increase the scrollbar scale (instead of using a rectangle
with full diagram dimensions), position of this small rect is still
updated during BPMN DI update. this solves problems like ctrl+a
and leaves move / create logic untouched.