feat(event+gateway/decorators): employ decorators
This commit aligns the decoration handling for events and gateways with
the one employed for service tasks. Instead of attaching the decoration
feature to `IAddFeature` it is now refactored to an own concern for both
elements.

Additionally it adds some minor fixes to some of our features.

- employ IDecorateFeature for events
- employ IDecorateFeature for gateways

- fix model sync bug on adding event definition
 (AAEventDefinitionFeature)
- fix recursive lookup of business objects (BusinessObjectUtil)
- change preferences to show XOR marker per default (Bpmn2Preferences)
- update connection check to switch on EType of connections instead
  of enum (ConnectionOperations)
- lookup actual EType before creating bpmn2 model elements
  (ABpmn2CreateFeature, AMorphNodeFeature)

Related to CAM-308, CAM-286, CAM-306, CAM-312, CAM-310, CAM-1415