feat(tests): add abstract integration test base
This commit adds a new integration test case class to the modeler test
suite.

The AbstractIntegrationTest may be used to open, manipulate and save
a diagram in a modeler. It resambles the actual editor behavior as much
as possible.

A test case may use the API methods #open, #execute and #save to
interact
with the opened editor.

A simple integration test built on top of the new foundation is the
RoundtripTest.

Closes CAM-1496